Legendre

FiniteHilbertTransform.LegendreFHT — Type

LegendreFHT

Type representing Legendre Finite Hilbert Transform (FHT) parameters.

This struct stores the necessary parameters for performing a Finite Hilbert Transform using Legendre functions. It implements the AbstractFHT interface and provides the required data structures and computations for Legendre FHT.

Fields:

name::String: FHT name (default: "Legendre").

Ku::Int64: Number of sample points.

tabu::Array{Float64,1}: Array containing u values (sampling points).

tabw::Array{Float64,1}: Array containing w values (weights at sampling points).

tabP::Matrix{Float64}: Matrix containing P_k(u) values (Ku x Ku).

tabc::Vector{Float64}: Vector containing prefactor values at each sampling point.

tabPLeg::Array{ComplexF64,1}: Static container for tabPLeg (Legendre functions of the first kind).

tabQLeg::Array{ComplexF64,1}: Static container for tabQLeg (Hilbert-transformed Legendre functions).

tabDLeg::Array{ComplexF64,1}: Static container for tabDLeg (Derivatives of Legendre functions).

Example:

Ku = 100
tabu = collect(-1.0:2/(Ku-1):1.0)
tabw = compute_weights(tabu) # Compute weights for Legendre functions
tabP = compute_legendre_matrix(tabu) # Compute Legendre functions matrix
tabc = compute_prefactors(tabu) # Compute Legendre prefactors

FHT = LegendreFHT("Legendre", Ku, tabu, tabw, tabP, tabc, zeros(ComplexF64, Ku), zeros(ComplexF64, Ku), zeros(ComplexF64, Ku))

FiniteHilbertTransform.GettabD! — Method

fill struct_tabLeg at a given complex frequency for the integration being considered

integration style selection is automatic: if you want to specify a type, call out to the specific integration method.

FiniteHilbertTransform.tabQLeg! — Function

tabQLeg!(omg::ComplexF64, val_0::ComplexF64, val_1::ComplexF64, tabQLeg::Array{ComplexF64,1})

Precompute Hilbert-transformed Legendre functions for a given complex frequency.

Arguments:

omg::ComplexF64: Complex frequency. (ATTENTION: Must be complex.)

val_0::ComplexF64: Initial value for k = 0. (ATTENTION: Must be complex.)

val_1::ComplexF64: Initial value for k = 1. (ATTENTION: Must be complex.)

tabQLeg::Array{ComplexF64,1}: Container to store the precomputed Hilbert-transformed Legendre functions.

Details:

This function uses different recurrence relations based on the location of the complex frequency omg. If omg is sufficiently close to the real line [-1,1], it employs an upward recurrence. Otherwise, if omg is far away from the real line [-1,1], it uses a backward recurrence. The transition between these regimes is determined dynamically.

The transition from the two regimes follows from the thesis Stable Implementation of Three-Term Recurrence Relations, Pascal Frederik Heiter, June, 2010 https://www.uni-ulm.de/fileadmin/websiteuniulm/mawi.inst.070/funken/bachelorarbeiten/bachelorthesis_pfh.pdf

Example:

omg = 1.0 + 2.0im
val_0 = 1.0 + 1.0im
val_1 = 2.0 + 2.0im
Ku = 10
tabQLeg = zeros(ComplexF64, Ku)
tabQLeg!(omg, val_0, val_1, tabQLeg)

Chebyshev

FiniteHilbertTransform.ChebyshevFHT — Type

ChebyshevFHT

FiniteHilbertTransform.GettabD! — Method

fill fht at a given complex frequency for the integration being considered

integration style selection is automatic: if you want to specify a type, call out to the specific integration method.